[GH-507] Change GetConnection Struct (#507)
